requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
db-dtypes==1.1.1
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.10.0
google-cloud-bigquery-storage==2.19.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.7
packaging==23.1
pandas==2.0.1
pandas-gbq==0.19.2
proto-plus==1.22.2
protobuf==4.23.1
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydata-google-auth==1.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2
